requirements.txt
streamlit
pandas
python-docx
plotly
sqlalchemy
